// frontend/src/types/schemas.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import * as zod from 'zod';

export const registerBodyFirstNameMin = 2;
export const registerBodyFirstNameMax = 2147483647;

export const registerBodyLastNameMin = 2;
export const registerBodyLastNameMax = 2147483647;

export const registerBodyPasswordMin = 6;
export const registerBodyPasswordMax = 2147483647;

export const registerBodyUsernameMin = 3;
export const registerBodyUsernameMax = 2147483647;



export const RegisterBody = zod.object({
  "firstName": zod.string().min(registerBodyFirstNameMin).max(registerBodyFirstNameMax),
  "lastName": zod.string().min(registerBodyLastNameMin).max(registerBodyLastNameMax),
  "email": zod.string(),
  "password": zod.string().min(registerBodyPasswordMin).max(registerBodyPasswordMax),
  "username": zod.string().min(registerBodyUsernameMin).max(registerBodyUsernameMax),
  "role": zod.enum(['STUDENT', 'TEACHER', 'ADMIN'])
})

export const RegisterResponse = zod.unknown()